New updates for a _getline function using MAX_ARG macro

+++ holberton.h
@@ -11,7 +11,7 @@ int _strcmp(char *, char *);
 char *_strcat(char *dest, char *src);
 void my_exit(char *exit_status, char **command);
 int _atoi(char *);
-short _getline(char **command, size_t *size, int fd);
+short _getline(char **command, int *size, int fd);
 void (*signal(int, void (*)(int)))(int);
 
 #endif /* HOLBERTON_H */
